Contract-first budget API and structured persistence (#17)
Refactor budget API and persistence around a checked-in /api/v1 contract.

- replace map-based report JSON with typed budget snapshots and normalized transaction rows
- add structured Flyway V4 schema for reports, snapshots, transactions, and import audits
- update React data flow to use dashboard snapshots, server-side analytics, and paginated transactions
- add OpenAPI contract, repository tests, and fake CSV API integration coverage

Closes #16

+++ src/main/java/com/budget/application/importing/ImportSummary.java
@@ -1,13 +1,14 @@
 package com.budget.application.importing;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record ImportSummary(
         String status,
         List<Integer> years,
         int transactions,
-        double income,
-        double spend,
+        BigDecimal income,
+        BigDecimal spend,
         String message
 ) {
 }
